[VPlan] Move lowering transforms to VPlanLowering.cpp (NFC) (#209885)
Following up to https://github.com/llvm/llvm-project/pull/209883,
 move transformations related to lowering/preparing for execution
to VPlanLowering.cpp

This moves materialize*, convertToConcreteRecipes,
dissolveLoopRegions, expandBranchOnTwoConds,
replaceWideCanonicalIVWithWideIV, expandSCEVsToVPInstructions, 
addBranchWeightToMiddleTerminator and the alias-mask materialization helpers.

The createScalarIVSteps and scalarizeVPWidenPointerInduction are
promoted to vputils.

Depends on https://github.com/llvm/llvm-project/pull/209883
